Refactor effects
