+++ src/effects/when/cellTimingFunctions.ts
